+++ R/find_mad.R
@@ -1,6 +1,6 @@
 #' @title Identify outliers based on 3 MAD
 #'
-#' @description Identify outliers based on 3 median absolute deviations.
+#' @description Identify outliers based on 3 median absolute deviations (MAD).
 #'
 #' See: Leys, C., Ley, C., Klein, O., Bernard, P., & Licata, L. (2013).
 #' Detecting outliers: Do not use standard deviation around the mean,
@@ -15,9 +15,14 @@
 #' @param criteria How many MAD to use as threshold (similar to standard
 #' deviations)
 #' @param mad.scores Logical, whether to output robust z (MAD) scores (default)
-#'                   or raw scores.
+#'                   or raw scores. Defaults to `TRUE`.
 #' @keywords standardization normalization median MAD mean outliers
 #' @author Hugues Leduc, Charles-Étienne Lavoie, Rémi Thériault
+#' @return A list of dataframes of outliers per variable, with row
+#'         numbers, based on the MAD. When printed, provides the number
+#'         of outliers, selected variables, and any outlier flagged for
+#'         more than one variable. More information can be obtainned
+#'         by using the `attributes()` function around the generated object.
 #' @export
 #' @examples
 #' find_mad(
@@ -37,6 +42,40 @@
 #' @importFrom dplyr mutate %>% select ends_with across all_of
 #' if_any filter bind_rows count n
 
+#' @export
+print.find_mad <- function(x, ...) {
+  mad0 <- attr(x, "outlier_total")
+  duplicates.df <- attr(x, "outlier_multiple")
+  criteria <- attr(x, "criteria")
+  col.list <- attr(x, "col.list")
+  mad0.list <- attr(x, "outlier_list")
+
+  if (isTRUE(nrow(mad0) > 0)) {
+    cat(
+      nrow(mad0), "outlier(s) based on", criteria,
+      "median absolute deviations for variable(s): \n",
+        paste0(col.list, collapse = ", "), "\n\n"
+    )
+    if (nrow(duplicates.df) > 0) {
+      cat(
+        "The following participants were considered outliers ",
+        "for more than one variable: \n\n",
+        sep = ""
+      )
+      print(duplicates.df)
+      cat("\n")
+    }
+    cat("Outliers per variable: \n\n")
+    print(mad0.list)
+  } else {
+    cat(
+      "There were no outlier based on", criteria,
+      "median absolute deviations.\n\n"
+    )
+  }
+}
+
+#' @export
 find_mad <- function(data,
                      col.list,
                      ID = NULL,
@@ -46,7 +85,7 @@ find_mad <- function(data,
     data$ID <- rownames(data)
   }
   row.names(data) <- NULL
-  mad0 <- find_mad0(data, col.list, ID = ID, criteria = criteria, mad.scores = mad.scores)
+
   mad0.list <- lapply(col.list, function(x) {
     find_mad0(data,
               x,
@@ -57,6 +96,7 @@ find_mad <- function(data,
   })
   names(mad0.list) <- col.list
   mad0.list <- mad0.list[lapply(mad0.list, nrow) > 0]
+
   duplicates.df <- bind_rows(mad0.list) %>%
     select(where(~ !all(is.na(.x))))
   if (nrow(duplicates.df) > 0) {
@@ -71,29 +111,18 @@ find_mad <- function(data,
     duplicates.df <- duplicates.df %>%
       filter(n > 1)
   }
-  if (nrow(mad0) > 0) {
-    cat(
-      nrow(mad0), "outlier(s) based on", criteria,
-      "median absolute deviations for variable(s): \n",
-      paste0(col.list, ", "), "\n\n"
-    )
-    if (nrow(duplicates.df) > 0) {
-      cat(
-        "The following participants were considered outliers ",
-        "for more than one variable: \n\n",
-        sep = ""
-      )
-      print(duplicates.df)
-      cat("\n")
-    }
-    cat("Outliers per variable: \n\n")
-    mad0.list
-  } else {
-    cat(
-      "There were no outlier based on", criteria,
-      "median absolute deviations.\n\n"
-    )
-  }
+  mad0 <- find_mad0(data, col.list, ID = ID, criteria, mad.scores)
+
+  # Define attributes!
+  out <- mad0.list
+  class(out) <- "find_mad"
+  attr(out, "outlier_list") <- mad0.list
+  attr(out, "outlier_total") <- mad0
+  attr(out, "outlier_multiple") <- duplicates.df
+  attr(out, "criteria") <- criteria
+  attr(out, "col.list") <- col.list
+  out
+
 }
 
 find_mad0 <- function(data, col.list, ID = ID, criteria = 3, mad.scores) {
@@ -105,7 +134,7 @@ find_mad0 <- function(data, col.list, ID = ID, criteria = 3, mad.scores) {
     mutate(
       Row = row(.[1]),
       across(all_of(col.list), rempsyc::scale_mad,
-        .names = "{col}_mad"
+             .names = "{col}_mad"
       )
     ) %>%
     filter(if_any(ends_with("_mad"), ~ . > criteria | . < -criteria))