+++ R/nice_lm.R
@@ -16,6 +16,9 @@
 #' `lm` function for the models.
 #'
 #' @keywords moderation interaction regression
+#' @return A formatted dataframe of the specified lm model, with DV, IV, degrees
+#'         of freedom, regression coefficient, t-value, p-value, and the effect
+#'         size, the semi-partial correlation squared.
 #' @export
 #' @examples
 #' # Make and format model