+++ R/nice_lm_slopes.R
@@ -18,6 +18,9 @@
 #' function for the models.
 #'
 #' @keywords moderation interaction regression
+#' @return A formatted dataframe of the simple slopes of the specified lm model,
+#'         with DV, levels of IV, degrees of freedom, regression coefficient, t-value,
+#'         p-value, and the effect size, the semi-partial correlation squared.
 #' @export
 #' @examples
 #' # Make and format model