+++ R/nice_mod.R
@@ -21,6 +21,9 @@
 #' function for the models.
 #'
 #' @keywords moderation interaction regression
+#' @return A formatted dataframe of the specified lm model, with DV, IV, degrees
+#'         of freedom, regression coefficient, t-value, p-value, and the effect
+#'         size, the semi-partial correlation squared.
 #' @export
 #' @examples
 #' # Make the basic table