+++ R/nice_na.R
@@ -17,15 +17,32 @@
 #' a total of 4 missing data points out of 100 were observed,
 #' with no participant missing more than a single data point”).*
 #'
-#' See: Parent, M. C. (2013). Handling item-level missing
-#' data: Simpler is just as good. *The Counseling Psychologist*,
-#' *41*(4), 568-600. https://doi.org/10.1177%2F0011000012445176
-#'
 #' @param data The data frame.
 #' @param vars Variable (or lists of variables) to check for NAs.
 #' @param scales The scale names to check for NAs (single character string).
 #' @keywords missing values NA guidelines
+#' @return A dataframe, with:
+#'  - `var`: variables selected
+#'  - `items`: number of items for selected variables
+#'  - `na`: number of missing cell values for those variables (e.g., 2 missing
+#'  values for first participant + 2 missing values for second participant
+#'  = total of 4 missing values)
+#'  - `cells`: total number of cells (i.e., number of participants multiplied by
+#'  number of variables, `items`)
+#'  - `na_percent`: the percentage of missing values (number of missing cells,
+#'  `na`, divided by total number of cells, `cells`)
+#'  - `na_max`: The amount of missing values for the participant with the most
+#'  missing values for the selected variables
+#'  - `na_max_percent`: The amount of missing values for the participant with
+#'  the most missing values for the selected variables, in percentage
+#'  (i.e., `na_max` divided by the number of selected variables, `items`)
+#'  - `all_na`: the number of participants missing 100% of items for that scale
+#'  (the selected variables)
+#'
 #' @export
+#' @references Parent, M. C. (2013). Handling item-level missing
+#' data: Simpler is just as good. *The Counseling Psychologist*,
+#' *41*(4), 568-600. https://doi.org/10.1177%2F0011000012445176
 #' @examples
 #' # Use whole data frame
 #' nice_na(airquality)
@@ -44,22 +61,19 @@
 #'   c(sample(c(NA, 1:10), replace = TRUE), NA, NA, NA)
 #' }
 #' df <- data.frame(
+#'   ID = c("idz", NA),
 #'   scale1_Q1 = fun(), scale1_Q2 = fun(), scale1_Q3 = fun(),
 #'   scale2_Q1 = fun(), scale2_Q2 = fun(), scale2_Q3 = fun(),
 #'   scale3_Q1 = fun(), scale3_Q2 = fun(), scale3_Q3 = fun()
 #' )
 #'
 #' # One can list the scale names directly:
-#' nice_na(df, scales = c("scale1", "scale2", "scale3"))
+#' nice_na(df, scales = c("ID", "scale1", "scale2", "scale3"))
 #'
 #' @importFrom dplyr select all_of bind_rows summarize %>% first last
 
 nice_na <- function(data, vars, scales) {
   classes <- lapply(data, class)
-  if (any(!(classes %in% "numeric"))) {
-    warning("Some variables are not numeric. ",
-            "They are ignored for calculating the `all_na` column.")
-  }
   if (missing(vars) & missing(scales)) {
     vars.internal <- names(data)
   } else if (!missing(scales)) {
@@ -87,7 +101,9 @@ nice_na <- function(data, vars, scales) {
 }
 
 nice_na_internal <- function(data) {
+
   data %>%
+    ungroup() %>%
     summarize(
       var = paste0(first(names(.)), ":", last(names(.))),
       items = ncol(.),
@@ -95,6 +111,7 @@ nice_na_internal <- function(data) {
       cells = prod(dim(.)),
       na_percent = round(na / cells * 100, 2),
       na_max = max(rowSums(is.na(.))),
-      na_max_percent = round(na_max / ncol(.) * 100, 2),
-      all_na = sum(is.na(rowMeans(select(., where(is.numeric)), na.rm = TRUE))))
+      na_max_percent = round(na_max / ncol(.) * 100, 2)) %>%
+    mutate(all_na = sum(apply(data, 1, function(x) sum(is.na(x))) == ncol(data)))
+
 }