+++ R/nice_randomize.R
@@ -11,6 +11,8 @@
 #' @param col.names The desired additional column names for a runsheet.
 #'
 #' @keywords randomization conditions random allocation experimental design
+#' @return A dataframe, with particiapnt ID and randomized condition, based
+#'         on selected design.
 #' @export
 #' @examples
 #' # Specify design, number of conditions, number of